add video4linux support
Add support for video4linux webcams using the new Linux driver wrapper.
This support is still a bit hackish since the code assumes all ldev devices
will be webcams.  This is okay for now since all ldev devices WILL BE webcams,
but when new device classes are added this will be a problem.  Hopefully by
then, the driver will expose more information about the device.

Thanks goes to Lachlan Michael for helping to test this.