add support for SW_RADIO switches
Further research and experimentation shows that the SW_RADIO aspect of this
device does correspond to the hardware RF kill switch.  So the right thing to
do is to teach input_test_switch() and other related code how to cope with
SW_RADIO and button.type "radio".  The attached patch does exactly that.  It's
been tested with Linux kernel version 2.6.23.9-85.fc8 and thinkpad_acpi module
version 0.18-20071203.

SW_RADIO was added to the Linux kernel "input.h" header in version 2.6.22, but
presumable we want HAL to build against earlier kernel versions too.  The patch
uses ifdefs to check that SW_RADIO exists and omits the new code if it does
not.