fix /sbin/umount.hal to avoid always returning error if it succeeded
Zdenek Prikryl <zprikryl@redhat.com> reported this bug on the hal
mailing list (see below). The fix was simple; the return code type is
INT32, not UINT32.

Hello,
according this page:
http://lists.freedesktop.org/archives/hal/2007-March/007738.html, umount.hal
should return correct return code. So, if I put a cd into a cdrom, then an
automounter will mount it. In mtab in cdrom record is uhelper=hal and uid with
correct value. If I want to umount it in a console as a user (uid of the user ==
uid in mtab), then everything is fine, the cdrom is umounted. But umount itself
returns 1 as a return code, because /sbin/umount.hal returns 1. According the
upstream, it should be fixed (page above),  but in f8 it isn't fixed. More
precisely, the patch is applied but it returns always 1 as the return code.

Any idea where is a problem?