use g_timeout_add_seconds() where appropriate if glib 2.14 is available
As you probably know already, glib 2.14 introduced variants of the
g_timeout_add() functions that operate at whole second granularity. This
allows timers to be grouped, which results in a more power and CPU
efficient behavior.

Attached is a patch that adds a configure check for GLib 2.14 and
changes some code (notably in the Linux specific parts) to use the new
functions when possible.