properly serialize the calls to 'hal-acl-tool --reconfigure'
This fixes a long standing bug where ACL's were sometimes wrong when
VT switching and when resuming from STR or STD. The problem was that
we would get a bunch of events from ConsoleKit at once, say

 1. Session1 -> Inactive
 2. Session2 -> Active
 3. Session2 -> Inactive
 4. Session1 -> Active

So for events 1-4 we would spawn 'hal-acl-tool --reconfigure' with a
carefully prepared environment reflecting the ConsoleKit state. All
four copies would start at once and race to get the lock. Which means
that it was undefined what other they execute in.

It's very easy to reproduce this; just try running

 # chvt 1; chvt 7; sleep 1; getfacl /dev/snd/pcmC0D0p
 (or some other device node)

a few times and observe that sometimes you wouldn't get an ACL.

This patch serializes the events and fixes this bug.