compute udis that are unique in the gdl and tdl
Introduces hald_compute_udi which computes a udi thats unique in both the
tdl and gdl and uses it everywhere relvent. This is necessary when we're
processing events in parallel as otherwise its possible to get a race condition
with two devices added with the same name. As part of this patch,
hal_util_compute_udi grwos a varargs version and no longer makes the computer
udi unique. I've also added a hal_util_make_udi_unique for use in those places
where hal_util_compute_udi was only used to check a udi was unique when
inserting into the gdl.