move acl-list from /var/lib/hal/ to /var/run/hal
List of applied ACLs can get corrupted, and that prevents hal-acl-tool
from ever touching it again and fixing. If that was due to a crash, etc,
it is not valid any longer anyways. In that case probably it would make
sense to relocate /var/lib/hal/acl-list into /var/run/hal, and let it
be removed by operating system startup scripts.

Simple change that moves acl-list to /var/run from /var/lib, as FHS suggests:
> This directory contains system information data describing the system
> since it was booted. Files under this directory must be cleared
> (removed or truncated as appropriate) at the beginning of the boot
> process.

Note that packagers will probably want to move the acl list in post
installation scriptlet when updating.