add iMac (x86 linux) backlight addon
Added iMac (x86 linux) backlight addon.

On my way to control the hardcore radioactive backlight of the newer
iMac 24" aluminum models some older code from the mactel guys appeared
to work (while the brightness level range lacks on the 24" but is good
on the 20" ones, compared to Mac OS).

I glued a patch which exposes the control like the other Macbook and
Macbook Pro addons. People reported that it apparently also worked on
other iMac models which I added to the corresponding fdi.

With the lately upstreamed hid patches to support the new aluminum apple
keyboard (really nice!) one can control the backlight of the display
nicely.