fix dock station status detection on undock operation
When the user triggers an undock operation, the kernel signals a 'undock
triggered' event. To that time, the 'docked' sysfs property must not
necessarily be updated already, because the undock may (although unlikely)
fail. So add a callback to refresh the dock status until kernel flags tell
us that the undock operation has completed.