autofix primary (ACPI) batteries which charge with a rate > 50 W
This patch try to detect automatically broken batteries which tell that they
charge a primary (Laptop) battery with a chargerate higher than 50 W
(50.000 mW), which is IMO in 100% of the cases wrong for Laptop batteries.
Maybe we can also assume a much lower level as criterion.

To get valid remaining time values this patch set in such cases
battery.remaining_time.calculate_per_time=true and calculate the remaining
time based on time and the chargelevel.