fixed usage of info.subsystem and linux.subsystem
Fixed the Linux code to correct usage of the info.subsystem
property. This property is mandatory for all devices, so set
it always if it wasn't set in the .add handler. Set also
always linux.subsystem since it can differ from info.subsystem
and this key always point to the correct linux kernel
subsystem. It's also needed to use/get the correct handlers.

This should also fix fd.o #15679