don't leak file descriptor on umount
I think this is a leaked file descriptor from hal.  Hal opens the lock
file for write and then fails to call fcntl(fd,F_SETFD, FD_CLOSEXEC)

When the confined mount program runs, the SELinux kernel notices the open file
descriptor, checks the domain to see if it has access, then closes it with the
error.