make hal start a lot faster
Hal is a significant part of the boot time of (optimized) systems,
so optimizing hal to start faster is certainly worthwhile.

The patch below cuts the start time of hal by about one third,
by reducing the number of calls to hal_device_get_udi() from
well over 700.000 to below 100.000.