fixed Lock() handling
Fixed Lock() handling.

The current code store the lock information for a lock owner into a
hash-table with these info: key=service_name of the lock owner and
as value the device which was locked. Unfortunately this don't work
if a service request more than one lock, because the old one get
overwritten in the hash table. Use instead of the device a GSList
with devices. This should solve

If the lock owner dies, remove the locks of the old_service_name and
not of the new_service_name (otherwise the locks get never released).