fix issue on 64bit big endian architectures.
On 64bit big endian architectures, PowerPC64 in my case, off_t is 64bit. Now
rule.jump_position is only 32bit. the pad32_write function call in
set_jump_position only saves 32bit. On little endian architectures this are
the "correct" bits, but not on 64bit architectures which are big endian. The
solution is to type cast off_t to a 32bit value and save this one instead of
the original variable.