allow to cross compile hal with host and target having different input.h
Allow to cross compile hal with host and target having different input.h.

When the Makefile is picking /usr/include/linux/input.h and libc kernel
headers are newer than the one for target I will get a nice compile error
that the new keynames are not known.

The attached patch allows to inform hal where my input.h resides and allows
to cross compile it.