added new properties to mount alternative fstypes as e.g. ntfs-3g
Added new properties to mount alternative fstypes as e.g. ntfs-3g
to the spec:
- volume.fstype.alternative (string list)
  --> e.g. ntfs-3g, ntfs-fuse
- volume.mount.[alternative].valid_options
  --> e.g. volume.mount.ntfs-3g.valid_options={'ro','nodev',...}
- volume.unmount.[alternative].valid_options
  --> e.g. volume.unmount.ntfs-3g.valid_options={'lazy',...}

These new keys need some changes in the existing mount tools (as
e.g. under KDE, GNOME) to use the new features. To take advantage
of the new keys, the related mount handler (or filesystem) as e.g.
ntfs-3g need to deliver a own FDI-file which set the needed keys.

To use the new feature with the Mount() method of the
org.freedesktop.Hal.Device.Volume interface the caller need to
pass the alternative fstype to the fstype parameter of the method.
Otherwise the original fstype (detected by HAL) is used to mount.
If a alternative fstype is passed to the method, only the in
volume.mount.[alternative].valid_options listed options are allowed.