fix handling of Device Mapper devices for LUKS
HAL has currently a massive problem with Device Mapper devices,
especially on coldplug, and parallel device handling .

Since dm-devices are virtual devices and have no parent we can
never say (when HAL process the hotplug queue) which the related
parent storage device (e.g. /dev/sdb1) is and when it will be
available.

Changed HAL to identify dm-devices on hotplug/coldplug and to
mark them as such in the HotplugEvent with is_dm_device=TRUE.
When the hotplug queue get processed, delay all dm-device as long
as there are still non-dm block devices in the queue to be sure
that all block devices/volumes are available when the Device
Mapper block devices get handled.