increase hal helper timeout to fix slow CD-ROMs
in https://launchpad.net/bugs/218834 we got a report that some CD-ROMs
are too slow to properly detect everything within the standard 10
seconds hal helper timeout:

[6188]: 22:22:56.009 [D] probe-volume.c:379: Doing probe-volume for
/dev/scd0
[...]
[6182]: 22:22:58.002 [I] addon-storage.c:363: ... device /dev/scd0 is
not locked on HAL
[6188]: 22:23:05.558 [D] probe-volume.c:542: volume.disc.capacity =
2384855040
[...]
[6188]: 22:23:06.406 [E] probe-volume.c:302: Timed out probing
/dev/scd0 - broken device driver?

(NB the timestamps). Since the default 10 seconds seems to be pretty
much arbitrary, I just raised it to 20 for now.

Fixes bz #13009.