add ACL support to removable block devices
This adds support for adding ACLs to removable block devices (such as
USB keys) on insertion. This is useful for allowing users to run mkfs
or fsck on their USB keys.

Please note that the patch does *not* enable it by default, that's up
to distros or individual users (gnome-polkit-authorization), but it
provides a standard flip to configure this behaviour.