allow child timeout to be set so hal starts on big servers
When a server has a _lot_ of disks, hald fails to
startup as it spends more than 250 seconds probing.
We don't want to hardocde this number any higher, as
this may leave haldaemon hung for a long time if there
is an error. Add as a command line argument that can be
appended in the init script if this is required.