don't add input capabilities to tablet PNP device
If hal detects that a PNP device is associated with a tablet PC it
adds the input capability to that device. This is wrong - the input
device is the serial device whose parent is this PNP device. Adding
the input capability just results in X attempting to add the PNP
device as an input device, despite it not being one and having no
associated driver.

FDI files for the appropriate driver can add the capabilities to the
serial port device, so there's no need to have the core add these
misleading ones. Simply remove them.