fix hotplug_event_process_queue() to get reprobe working
Fixed hotplug_event_process_queue() to get reprobe working.
Added new function to check if there is still a REMOVE/ADD event
for the same device (sysfs_path) in the queue of the running
hotplug events. Wait for those events to prevent trouble if
e.g. a HOTPLUG_ACTION_REMOVE is still running and a
HOTPLUG_ACTION_ADD is the next event.