don't remove scsi_host if there are still targets using it
When a scsi device is removed, the host it is connected to is also
removed by missing_scsi_host().  However, it is possible that other
targets still exist that are connected to the same host.  In that case,
the scsi_host should not be removed.  We check if the scsi_host has
more than one child, and if so it is not removed.

fd.o#20712